Fixed transactioning. Added marketGroups.

+++ src/main/java/gavinh/eve/data/ItemTypeRepository.java
@@ -2,6 +2,6 @@ package gavinh.eve.data;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemTypeRepository extends CrudRepository<ItemType, Long> {
+public interface ItemTypeRepository extends CrudRepository<ItemType, Integer> {
     
 }
